Migrate UnusedSchemaExplorer
